LPS-78459 Rename SourceChecksSuppressions -> SourceFormatterSuppressions
